model method check params

+++ app/models/issue.rb
@@ -78,22 +78,25 @@ class Issue < Note
   # array to current issue.
   # Then delete those issues without Evidence.
   # Returns the number of issues affected.
-  def combine(issue_ids)
-    # assert current id is not there
+  def merge(issue_ids)
+    merged = 0
     issue_ids = [issue_ids] if issue_ids.is_a?(Integer)
+
+    # check all specified ids are integers
+    return merged unless issue_ids.all?{|i| i.is_a? Integer}
+
+    # assert current id is not there
     issue_ids -= [id]
 
-    combined = 0
-
-    # combine
+    # merge
     if issue_ids.any?
       self.transaction do
         Evidence.where(issue_id: issue_ids).update_all(issue_id: id)
-        combined = Issue.where(id: issue_ids).destroy_all.count
+        merged = Issue.where(id: issue_ids).destroy_all.count
       end
     end
 
-    combined
+    merged
   end
 
 end
